Remove non-blocking mailbox queue implementation
