+++ encon-config/src/test/resources/connector.yml
@@ -11,8 +11,6 @@ defaults:
   distribution-flags:
     - MAP_TAG
     - BIG_CREATION
-  mailbox:
-    blocking: false
   server:
     boss-threads: 2
     worker-threads: 4
@@ -38,9 +36,7 @@ nodes:
       - BIG_CREATION
     mailboxes:
       - name: net_kernel
-        blocking: false
       - name: another
-        blocking: true
       - name: another_one
     server:
       port: 8971
@@ -52,4 +48,3 @@ nodes:
     cookie: popa
     mailboxes:
       - name: net_kernel
-        blocking: false
