+++ encon-handler/README.md
@@ -14,7 +14,7 @@ First of all, add dependency:
   <dependency>
     <groupId>io.appulse.encon</groupId>
     <artifactId>encon-handler</artifactId>
-    <version>1.4.0</version>
+    <version>1.5.0</version>
   </dependency>
   ...
 </dependencies>
@@ -23,7 +23,7 @@ First of all, add dependency:
 **Gradle**:
 
 ```groovy
-compile 'io.appulse.encon.java:encon-handler:1.4.0'
+compile 'io.appulse.encon.java:encon-handler:1.5.0'
 ```
 
 ### Basics
