+++ pom.xml
@@ -24,7 +24,7 @@ limitations under the License.
 
   <groupId>io.appulse.encon</groupId>
   <artifactId>encon-parent</artifactId>
-  <version>1.4.0</version>
+  <version>1.5.0</version>
   <packaging>pom</packaging>
 
   <properties>
@@ -71,7 +71,7 @@ limitations under the License.
     <url>https://github.com/appulse-projects/encon-java</url>
     <connection>scm:git:https://github.com/appulse-projects/encon-java.git</connection>
     <developerConnection>scm:git:https://github.com/appulse-projects/encon-java.git</developerConnection>
-    <tag>1.4.0</tag>
+    <tag>1.5.0</tag>
   </scm>
 
   <distributionManagement>
